+++ src/doubles/java/org/squiddev/cobalt/lib/doubles/Bignum.java
@@ -22,7 +22,7 @@
  * SOFTWARE.
  */
 
-package org.squiddev.cobalt.lib.fmt;
+package org.squiddev.cobalt.lib.doubles;
 
 import org.checkerframework.checker.signedness.qual.Signed;
 import org.checkerframework.checker.signedness.qual.Unsigned;
